+++ SpringBootBackend/src/main/java/com/example/demo/Repository/GradoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.Entity.Grado;
 
 @Repository
-public interface GradoRepository extends JpaRepository<Grado, Integer> {
+public interface GradoRepository extends JpaRepository<Grado, Long> {
 
 }
