+++ SpringBootBackend/src/main/java/com/example/demo/Repository/GradoXNivelRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.Entity.GradoXNivel;
 
 @Repository
-public interface GradoXNivelRepository extends JpaRepository<GradoXNivel, Integer> {
+public interface GradoXNivelRepository extends JpaRepository<GradoXNivel, Long> {
 
 
 }
